Slightly reduced memory footprint
Note: storing impulse times for later processing obviously uses up a lot of memory (40b), but there are advantages to doing it this way. When receiving data, the number of required cycles is reduced to
a bare minimum and in case of a failure, examining the array can be helpful.